+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+### Added
+- Header option: `x-context`
+- Parameter `deliveryTagContextList` in rewind exceptions for transfer message context
 
 ## [3.1.0] - 2022-05-16
 ### Added